added functionalty to imageUpload
