+++ app/imageUpload/imageController.js
@@ -4,25 +4,41 @@ var fs = require('fs');
 var imgur = require('imgur');
 var multer = require('multer');
 
+var upload = multer({
+  dest: 'uploads/',
+  limits: {
+    fileSize: 10000000 //20MB
+  },
+  fileFilter: function (req, file, cb) {
+    cb(null, file.mimetype.slice(0,6) === 'image/');
+  }
+});
+
+var uploadMulter = upload.single('image')
+
 module.exports.img = function (req, res, next) {
-  imgurUpload(req);
-  next();
+  uploadMulter(req, res, function (err) {
+    if (err) {
+      console.log(err)
+    }
+    imgurUpload(req, function (err, result) {
+      req.imgur = result;
+      next();
+    });
+  })
 };
 
-function imgurUpload(req) {
-  req.body.file = req.file;
-  if (req.body.file) {
+function imgurUpload(req, cb) {
+  if (req.file) {
     imgur
-      .uploadFile(req.body.file.path)
-      .then(function (json) {
-        fs.unlink(req.body.file.path, function () {
-          req.body.file.imgurUrl = json.data.link;
-          console.log(req.body)
-          return req.body
+      .uploadFile(req.file.path)
+      .then(function (result) {
+
+        fs.unlink(req.file.path, function (err) {
+          if (err) throw err;
+          cb(err, result);
         });
       })
-      .catch(function (err) {
-        res.send(err);
-      });
+      .catch(cb);
   }
 }
